Allow get_setting to choose specific sub_ids
Fix get_setting() and prep_data() to allow user to specify whether
they want to retrieve the entire subject ids dict or only a specific
group key.

prep_data() was failing to complete the check_datasizes() step because
it was trying to use the entire subject id dict obtained from
get_setting. To fix this, get_setting() was changed to allow for choosing
a specfic subject id group, and prep_data() was changed to retrieve
'all' subject ids.