Fix small script settings-related issues
In the process of testing my code on Sherlock, I made a few piecemeal
changes to script setting related code:

- Change 'w' (write only) to 'r+' for read-and-write access when
saving the obtained datasize to script_settings.json when calling
prep_data()

- Move most code within get_settings() inside try/except statement to
clarify whether any kwarg leads to an exception. Change exception
printed message to be more descriptive of which argument was used
upon triggering the exception for debugging convenience.

- Add and improve formatting of normal print messages when running
script_setup.py.

These changes were made in the hope of clearing up errors while
trying to run my code with our real dataset on Sherlock. It is not
clear at the time of this commit whether they will significantly
fix things, but this issues appeared related and the revisions seemed
sensible.