Fix json updating to overwrite rather than append
Add json writing feature to get_setting() so that updating
script_settings.json with new data will overwrite instead of append the
final json file.

When running prep_data(), attempting to retrieve the global datasize and
save it again to script_settings.json seems to result in the
json file unexpectedly containing two top-level dictionaries: one dict
with the original data before attempting to save to
['Parameters']['datasize'], and a second dict containing the same info
from the first except with the new 4d integer list under the 'datasize'
key. To fix this:

- Relocate the code that handled the json file from prep_data() to
get_setting() as a new feature that allows you to simply supply
the new datasize to the 'save_datasize' kwarg.
- Add step where the dict containing the json's contents are first
updated with the new datasize before the json file is opened again to
write to.
- Change the with...open() argument to 'wt' instead of 'rt+' (whose
purpose I've forgotten already) when opening the json file to be updated
with the new contents so that the json is overwritten instead of
appended.

These changes should now result in allowing for script_settings.json to
be directly updated with a new datasize list or tuple, and for the
update to be represented inside the json file as a single modified
top-level dict rather than the old top-level dict followed by a new
top-level dict.

I also changed script_setup.py and tools.get_setting() to now allow
users to define an alternative filename for their script settings json
file--the default name is still 'script_settings.json'.