Add .nii.gz option to save_data
To visualize my data, I need to provide it the nilearn visualization
funcitons with the computed results that have been converted to either a
nibabel Nifti1Image object or a Nifti format .nii.gz file. I already
have a save_data() function that is supposed to provide a single
interface to saving either .npy or .npz Numpy format files flexibly.
I add .nii.gz and so that, in addition to the desired filename to save
output to, it can either accept:

- the required Nifti1Image object containing computed
results, or

- the computed results array, the affine coordiantes array, and the
3d volume datasize tuple or list.

I have not tested this function yet, but I am committing my changes now
before buildling on this feature and finally testing it. I will make any
corrections in future commits.