Change get_setting to save other parameters
Change get_setting() kwarg 'save_datasize' to 'save_param' and allow it
to accept a tuple or list of the desired parameter subsection and
the data that should be saved.

get_setting() previously had a specific kwarg argument for saving
the datasize called 'save_datasize' because it was the only parameter
that needed to be saved to script_settings.json at the time. Now that
I have decided to store the affine coordinate array as a nested list
in script_settings.json as well, I replaced the save_datasize keyword
argument with save_param instead of just adding a new keyword argument.
save_param accepts a two-item list or tuple where index 0 is the name
of the parameter and index 1 is the data to be saved. As a result,
get_setting() should be more flexible to any additional paramters that
I might decide need to be saved in the future.

In addition, I replaced code in prep_data() that saved datasize by
going through the process of opening script_settings.json using the
'with' statement, with a single call of get_setting() using the
new 'save_param' kwarg.