Add numpy array handling to get_setting
Change get_setting() to convert numpy arrays as nested lists when
saving to, and convert nested lists back into numpy arrays when
retrieving from script_settings.json.

Now that I am changing get_setting() to handle saving and retrieving
affine coordinate array, it needed to handle converting this array
into a nested list and back again. When saving a numpy array,
the is converted into a list if it its type is numpy.ndarray. When
retrieving a list, if the parameter is 'affine' (this may be
expanded into a list of qualified parameters if add more in the future),
then the retrieved list is converted in a numpy.array and returned;
otherwise the unaltered list is returned. This change should allow
me to seamlessly handle affine coordinate arrays within
script_settings.json and get_setting().